add option to start without ssl
