+++ ktor-starter/src/jvmMain/kotlin/eu/vendeli/ktor/starter/KtorServer.kt
@@ -22,42 +22,45 @@ fun serveWebhook(wait: Boolean = true, serverBuilder: ServerBuilder.() -> Unit =
     val cfg = ServerBuilder().apply(serverBuilder)
     val serverCfg = cfg.server ?: EnvConfiguration
 
-    val keystoreFile = File(serverCfg.KEYSTORE_PATH)
-
-    val keystore = KeyStore.getInstance("JKS")
-    if (keystoreFile.exists()) {
-        keystore.load(keystoreFile.inputStream(), serverCfg.KEYSTORE_PASSWORD)
-    } else {
-        keystore.load(null, null)
-        val pemPrivateKeyFile = File(serverCfg.PEM_PRIVATE_KEY_PATH).also {
-            if (!it.exists()) throw IllegalStateException("PEM_PRIVATE_KEY file not found")
-        }
-        val chainFile = File(serverCfg.PEM_CHAIN_PATH).also {
-            if (!it.exists()) throw IllegalStateException("PEM_CHAIN_PATH file not found")
-        }
-
-        val pemPrivateKey: PrivateKey = PemUtils.loadPrivateKey(pemPrivateKeyFile.inputStream())
-        val pemChain = PemUtils.loadCertificate(chainFile.inputStream()).toTypedArray()
-
-        keystore.setKeyEntry(serverCfg.KEY_ALIAS, pemPrivateKey, serverCfg.PEM_PRIVATE_KEY, pemChain)
-        keystore.store(keystoreFile.outputStream(), serverCfg.KEYSTORE_PASSWORD)
-    }
-
     val environment = applicationEngineEnvironment {
         connector {
             host = serverCfg.HOST
             port = serverCfg.PORT
         }
-        sslConnector(
-            keyStore = keystore,
-            keyAlias = serverCfg.KEY_ALIAS,
-            keyStorePassword = { serverCfg.KEYSTORE_PASSWORD },
-            privateKeyPassword = { serverCfg.PEM_PRIVATE_KEY },
-        ) {
-            host = serverCfg.HOST
-            port = serverCfg.SSL_PORT
-            keyStorePath = keystoreFile
+        if (serverCfg.SSL_ON) {
+            val keystoreFile = File(serverCfg.KEYSTORE_PATH)
+            val keystore = KeyStore.getInstance("JKS")
+
+            if (keystoreFile.exists()) {
+                keystore.load(keystoreFile.inputStream(), serverCfg.KEYSTORE_PASSWORD)
+            } else {
+                keystore.load(null, null)
+                val pemPrivateKeyFile = File(serverCfg.PEM_PRIVATE_KEY_PATH).also {
+                    if (!it.exists()) throw IllegalStateException("PEM_PRIVATE_KEY file not found")
+                }
+                val chainFile = File(serverCfg.PEM_CHAIN_PATH).also {
+                    if (!it.exists()) throw IllegalStateException("PEM_CHAIN_PATH file not found")
+                }
+
+                val pemPrivateKey: PrivateKey = PemUtils.loadPrivateKey(pemPrivateKeyFile.inputStream())
+                val pemChain = PemUtils.loadCertificate(chainFile.inputStream()).toTypedArray()
+
+                keystore.setKeyEntry(serverCfg.KEY_ALIAS, pemPrivateKey, serverCfg.PEM_PRIVATE_KEY, pemChain)
+                keystore.store(keystoreFile.outputStream(), serverCfg.KEYSTORE_PASSWORD)
+            }
+
+            sslConnector(
+                keyStore = keystore,
+                keyAlias = serverCfg.KEY_ALIAS,
+                keyStorePassword = { serverCfg.KEYSTORE_PASSWORD },
+                privateKeyPassword = { serverCfg.PEM_PRIVATE_KEY },
+            ) {
+                host = serverCfg.HOST
+                port = serverCfg.SSL_PORT
+                keyStorePath = keystoreFile
+            }
         }
+
         modules.addAll(cfg.ktorModules)
         module {
             routing {
